requirements.txt
PyMySQL==1.1.0
python-dotenv==1.0.1
setuptools==69.5.1
wheel==0.43.0
